Renaming the file holding TokenizedBooleanFieldFilter and TokenizedValueListFilterFormBase to match the class names
