+++ Lombiq.Projections.csproj
@@ -119,11 +119,11 @@
     <Compile Include="Projections\Filters\ContainsContentFieldsFilter.cs" />
     <Compile Include="Projections\Filters\EmptySafeContentFieldsFilter.cs" />
     <Compile Include="Projections\Filters\TokenizedTaxonomyFieldTermsFilter.cs" />
-    <Compile Include="Projections\Filters\TokenizedBooleanFilter.cs" />
+    <Compile Include="Projections\Filters\TokenizedBooleanFieldFilter.cs" />
     <Compile Include="Projections\Filters\TokenizedContentItemIdsFilter.cs" />
     <Compile Include="Projections\Filters\TokenizedContentTypeFilter.cs" />
     <Compile Include="Projections\Forms\TokenizedStringValueListFilterForm.cs" />
-    <Compile Include="Projections\Forms\TokenizedValueListFilterForm.cs" />
+    <Compile Include="Projections\Forms\TokenizedValueListFilterFormBase.cs" />
     <Compile Include="Projections\Filters\TokenizedTaxonomyTermsFilter.cs" />
     <Compile Include="Projections\Forms\TokenizedTaxonomyFieldTermsFilterForm.cs" />
     <Compile Include="Projections\Forms\TokenizedTaxonomyTermsFilterForm.cs" />
